feat: change DB MySQL to Postgres, create new file .env, config value in file .env

+++ src/main/java/com/exe201/group1/psgp_be/dto/requests/ProductUpdateRequest.java
@@ -20,7 +20,7 @@ public class ProductUpdateRequest {
     String name;
     String description;
     String size;
-    BigDecimal price;
+    long price;
     Integer quantityInStock;
     String status;
     List<Integer> succulentIds;
